requirements.txt
insta-scrape==2.1.2
instaloader
torch
keras
opencv-python
tensorflow
tqdm
pillow
instagrapi
selenium
pandas
instareels
wget
apify_client
python-dotenv